[FIX] OPW 576010: selection widgets in editable lists should properly accept mouse selection
(Forward-port of 6.0 fix: revno: 1954 
revision-id: odo@openerp.com-20120726131647-tiy7m67ruo27n81j)

This seems to be a regression due to an upstream
change in GTK, as explained in the GTK issue:
 https://bugzilla.gnome.org/show_bug.cgi?id=317387

The original behavior can apparently be emulated
with the `changed` signal, as modified in GTK issue:
 https://bugzilla.gnome.org/show_bug.cgi?id=324282

Due to the various signal handlers we already have,
the `changed` event seems to propagate as we need,
but we failed to handle the signal chain properly
because it comes from a ComboBox at some point
and not a ComboBoxEntry.
This one-line patch to accept signals from
all ComboBox widgets seems to do the trick on
Ubuntu 11.10, and should be low-risk.